refactor/ quieries 파일 오타 수정 => queries
